Fix TypeErrors related to DataServiceProvider
DataServiceProvider is a static class - all of its methods are static
and it doesn't have an __INIT__ method. Therefore, it doesn't make sense
to instantiate it. Furthermore, the type hint for the data_provider
param of download_asset_files should be a class object
`Type[DataServiceProvider]` instead of an instance `DataServiceProvider`

Unfortunately the enforce-typing library does not seem to support object
class type hints because the download_asset_files method is called
during test_payer_market_flow with data_provider = Type[DataProviderMock]
but no TypeError occurs.